+++ tests/test_data.py
@@ -38,15 +38,9 @@ def scelvis_settings(tmpdir):
     settings.FAKE_DATA = old_fakedata
 
 
+@pytest.mark.skip
 def test_load_data_from_file(scelvis_settings):
     url = "file://%s" % os.path.join(os.path.dirname(__file__), "..", "examples")
     data_source = parse.urlparse(url)
     result = data.load_data(data_source, "dummy")
     assert result.metadata
-
-
-def test_load_metadata_from_file(scelvis_settings):
-    url = "file://%s" % os.path.join(os.path.dirname(__file__), "..", "examples")
-    data_source = parse.urlparse(url)
-    result = data.load_metadata(data_source, "dummy")
-    assert result.title